system-tray/requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
psutil>=5.9.0
portalocker>=2.8.2

